Runs promtool on files that lack "groups:" at top
This makes it so `promtool` runs only on files that were changed in the
PR instead of all of them. It also adds `oq` to the mix, which is a `jq`
wrapper that allows us to parse yaml and add the "groups:" key to the
root of each changed YAML document.

This is not likely an upstream change that will help the project from
which it's forked.